requirements.txt
Pillow==3.4.2
requests==2.11.1
Flask==0.11.1
Jinja2==2.8
Werkzeug==0.11.11
click==6.6
tensorflow==0.12.0
sklearn==0.0
six==1.10.0
scipy==0.18.1
scikit-learn==0.18.1
psycopg2==2.6.2
protobuf==3.1.0.post1
numpy==1.11.3
itsdangerous==0.24